add: delete with password stored in dotenv

+++ api/src/routes/api/v2/piece/index.js
@@ -1,5 +1,7 @@
 const router = require('express').Router();
 const jwt = require('jsonwebtoken');
+const config = require('../../../../config.js');
+
 const PieceStore = require('../../../../models/v2/PieceStore.js');
 const db = require('../../../../utils/db.js');
 
@@ -17,8 +19,16 @@ router.route('/')
     res.json({ status, piece });
   })
   .delete(async (req, res) => {
-    const pieces = PieceStore.deletePieces();
-    await db.deleteMongo();
-    res.json(pieces);
+    console.log('きたよー');
+    console.log(req.query);
+
+
+    if (req.query.keyword !== config.deletePass) {
+      res.sendStatus(204); // パスワードが違う場合
+    } else {
+      const pieces = PieceStore.deletePieces();
+      await db.deleteMongo();
+      res.json(pieces);
+    }
   });
 module.exports = router;
